pyproject.toml
[project]
authors = [
  {name = "Brian Dawn", email = "brian.t.dawn@gmail.com"},
]
dependencies = [
    "python-fasthtml>=0.3.2",
    "python-dotenv>=1.0.1",
    "mangum>=0.17.0",
]
description = "Add your description here"
name = "mess-with-fasthtml"
readme = "README.md"
requires-python = ">= 3.8"
version = "0.1.0"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.rye]
dev-dependencies = [
    "ruff>=0.5.7",
    "aws-cdk-lib>=2.151.0",
    "constructs>=10.0.0",
    "aws-cdk-aws-lambda-python-alpha>=2.151.0a0",
]
managed = true

[tool.rye.scripts]
cdk = "npx cdk@latest"
dev = "python src/mess_with_fasthtml/web/main.py"
fmt = "ruff format ."
lint = "ruff check . --fix"

[tool.ruff]
lint.extend-ignore = ["F403", "F405"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/mess_with_fasthtml"]
